+++ CHANGELOG.md
@@ -11,9 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added command line argument validation
 
 ### Changed
-- Changed **taskmanager.Create** return arguments, provided only error.
+- Changed **taskmanager.Read** return arguments, provided slices of tasks
+- Changed **taskmanager.Create** return arguments, provided only error
 
-## [0.0.1] 2020-07-26
+## [0.0.0] 2020-07-26
 ### Added
 - Added **taskmanager.Create** method
 - Added **simpletask.Task** and **simpletask.Comment** definitions